quitando código repetido en la eliminaación de carrito. Lo único es que aahora no se ven automáticamente los cambios aal agregar productos al caarrito, se debe espera a que la BD se elimine

+++ frontend/src/utils/carritoFunctions.js
@@ -24,9 +24,9 @@ export const handleRemoveItem = (idProducto, removeFromCart, setFilas) => {
 };
 
 export const handleClearCart = async (userId, dispatch, setFilas) => {
-  await deleteDB(userId);
   dispatch({ type: "CLEAR_CART" });
   setFilas([]);
+  if (userId) await deleteDB(userId);
 };
 
 export const deleteDatabase = deleteDB;
